render: address help-overlay review nits (#2550)
Two organizational nits from the Sonnet review + Opus recheck on #2622.

Split the visibility accessors out of the component header. The Sonnet
review flagged that `components/component_help_overlay.hpp` bundled
`C_HelpOverlayState` with the `IRPrefab::HelpOverlay::` accessor namespace,
diverging from the `C_WidgetTheme` precedent it cited — that precedent keeps
`components/` POD-only and puts the accessors in a sibling
(`widget_theme.hpp`). The Opus recheck upheld it and named the trap: folding
the accessors into the adoption surface cycles, because `help_overlay.hpp`
includes `systems/system_help_overlay.hpp` whose `create()` calls
`ensureStateSingleton()`. New `help_overlay_state.hpp` is the third header
both sides include instead. Function bodies move verbatim.

Fix the `textH` comment. It claimed the trailing newline on the last line
adds a row to `lineCount_`; `lineCount_` is incremented by `appendLine`, so
the `-1` is just the standard N-line height (N-1 inter-line steps at
`kGlyphStepY` = 12, plus one `kGlyphHeight` = 11 row). The math was right,
the explanation described a mechanism that isn't there.

Verified: render-verify IRShapeDebug 24/24 at max_delta 0 (byte-identity
holds), gui-verify --gui-test 5/5 (798 glyphs on the ON path, camera entry
described), 1371 engine tests, and all five adopter targets build clean.
