+++ engine/prefabs/irreden/render/components/component_help_overlay.hpp
@@ -1,11 +1,10 @@
 #ifndef COMPONENT_HELP_OVERLAY_H
 #define COMPONENT_HELP_OVERLAY_H
 
-#include <irreden/ir_entity.hpp>
-
 namespace IRComponents {
 
-// Singleton row holding the help overlay's visibility. World-scoped
+// Singleton row holding the help overlay's visibility, read via
+// `IRPrefab::HelpOverlay::` (`help_overlay_state.hpp`). World-scoped
 // mutable state lives on a singleton component rather than a header global
 // or a RenderManager field — see `.claude/rules/cpp-globals.md` and
 // `engine/prefabs/irreden/render/CLAUDE.md` §"Exposing system public API".
@@ -21,38 +20,5 @@ struct C_HelpOverlayState {
 };
 
 } // namespace IRComponents
-
-namespace IRPrefab::HelpOverlay {
-
-// Singleton accessor. Registration-time helper `ensureStateSingleton()`
-// below mirrors `widget_theme.hpp::ensureThemeSingleton()`: the lazy
-// `singleton<T>()` create is a `createEntity`, whose first call must run on
-// the main thread, so every consumer touches it from `create()` during
-// pipeline wiring rather than inside a first-frame `beginTick`.
-inline IRComponents::C_HelpOverlayState &state() {
-    return IREntity::singleton<IRComponents::C_HelpOverlayState>();
-}
-
-inline void ensureStateSingleton() {
-    state();
-}
-
-// Flip / query helpers so callers (the toggle command, a creation that wants
-// the overlay open at startup, a GUI test predicate) never spell the
-// singleton lookup themselves.
-inline void setVisible(bool visible) {
-    state().visible_ = visible;
-}
-
-inline void toggleVisible() {
-    auto &s = state();
-    s.visible_ = !s.visible_;
-}
-
-inline bool isVisible() {
-    return state().visible_;
-}
-
-} // namespace IRPrefab::HelpOverlay
 
 #endif /* COMPONENT_HELP_OVERLAY_H */
